Update todo list
Hi guys, see this list for my take on the remaining queue of tasks.
Any feedback or coding would be welcome.  Especially appreciated would be
trying this on a display that is connected to a motor (because it will be
a few days before I can do that)

Note: in the current main.c you will want to comment out configscreen_show()
and instead call mainscreen_show()

  // mainscreen_show();
  configscreen_show();