Add an accelerated fillrect/drawline routine
This wasn't a work item of ours but it looked like fun so I did it:
The adafruit folks have had a SSD1306 driver for a while.  That chip seems
to be _very_ similar to the oled controller on this board.  Even the frame
buffer layout is identical.  Therefore it was easy for me to port their
optimized line/fill code and install it in ugui.

This speeds up draw operations by about 2.3x.