Fix display pin bindings based on datasheets from web, see below...
For #9 I added lcd_read_data_16bits() and used it to read the device code.
First I had to fix some GPIO bindings (we had RESET and RD swapped).
But now I see that on mine it is a real ili9481.  We still need to add
something so that the bootscreen temporarily prints this data so @r0mko
can find out what sort of chip he has on his display.