Check in a prebuilt binary of the bootloader, see below
To prevent future possible problems (i.e. someone is building a buggy
version of bootloader).  We now default to using a prebuilt version of
the bootloader bin.  Currently corresponding to the osebike-bootloader-v1.0
tag in the bootloader project.

If you ever release a new bootloader, please create a new tag in that
project and add the bin to this repo.

attn @casainho & @lowperformer