My depdency checking for 850C header files works but has one ugly thing...
The .d files and .o files for common srcs are placed into the wrong directory
instead of being in .deps/_build they are in the 850C directory.  Because
I'm apparently doing something dumb with relative paths.  But since things
work and the build is SO MUCH faster (and no longer any bugs because of
forgetting to do a clean build) I'm checking this in for now and just
ignoring this wart.