WIP - everything now fits on SW102 again but (see below)
@casainho and @lowperformer
I need to fix a couple of things before it will run.  I'll do that
tomorrow night.  I decided to bite the bullet and do what I should have
done from the beginning - separate out the read only parts of screen
Fields (so they can be const and not use RAM) and move the small bits
of state that changes at runtime into a FieldRW struct.  Now everything
fits nicely in RAM on the SW102.

btw: Casainho if your large number of graphs are still causing 850C
to run out of RAM I highly encourage you to use a metaphor more like I had
originally for graphs - where the huge GraphData objects are handed out
of a cache (rather than 1:1 with each possible graph).  And those objects
are limited to only the graphs that are currently being shown.  Yes, this
means that when the user is not graphing <whatever> data for whatever is
not being stored.  But since multiple graphs can be on a screen I don't
think that limit is so bad.  And it saves immense amounts of RAM.

Btw - to track down what was burning RAM I used:

To view the map file I used the excellent http://www.sikorskiy.net/prj/amap/ tool:
~/packages/amap _build/nrf51822_sw102.map