Correct Assist level display on SW102 when set for Imperial
Assist level display was incorrect if SW102 was set to Imperial.
This is because the assist level was being converted to Miles Per hour because it was using the speed display.  So an assist level of 20 (KPH) was being converted to a display of 12 (MPH).
This change uses a lookup table to insert suitable values of Assist that when converted give the correct display.  e.g Assist 20 is converted to 33 (KPH) which after conversion displays as 20 (MPH)